Move JsonSerializerOptionsExtensions to Json.More
